+++ python/packages/aws-lambda-sdk/serverless_aws_lambda_sdk/internal_extension/wrapper.py
@@ -1,95 +1,75 @@
 from __future__ import annotations
-
 import importlib
-import logging
-import sys
 from os import environ
-from typing import List, Optional
+from typing import List
 
 from typing_extensions import Final, TypeAlias, TYPE_CHECKING
 
-from ..base import Handler
-from .base import Env
-from ..exceptions import HandlerNotFound
-
 from ..instrument import Instrumenter
-
-if TYPE_CHECKING:
-    from serverless_sdk.sdk.base import ServerlessSdk
-
-else:
-    ServerlessSdk: TypeAlias = "ServerlessSdk"
-
 
 __all__: Final[List[str]] = [
     "handler",
 ]
 
+if environ.get("_ORIGIN_HANDLER") is None:
+    raise Exception("Missing _ORIGIN_HANDLER environment variable")
 
-HANDLER_MODULE_BASENAME: Final[Optional[str]] = environ.get(Env.HANDLER_MODULE_BASENAME)
-HANDLER_BASENAME: Final[Optional[str]] = environ.get(Env.HANDLER_BASENAME)
-HANDLER_MODULE_DIR: Final[Optional[str]] = environ.get(Env.HANDLER_MODULE_DIR)
-HANDLER_FUNCTION_NAME: Final[Optional[str]] = environ.get(Env.HANDLER_FUNCTION_NAME)
+environ["_HANDLER"] = environ.get("_ORIGIN_HANDLER")
+del environ["_ORIGIN_HANDLER"]
 
-
-def _get_handler_function():
-    path_modified = False
-    if HANDLER_MODULE_DIR not in sys.path:
-        path_modified = True
-        sys.path.append(HANDLER_MODULE_DIR)
-    try:
-        module = importlib.import_module(HANDLER_MODULE_BASENAME)
-        return getattr(module, HANDLER_FUNCTION_NAME)
-    finally:
-        if path_modified:
-            sys.path.remove(HANDLER_MODULE_DIR)
+try:
+    from serverless_aws_lambda_sdk import serverlessSdk
+except ModuleNotFoundError:
+    from .. import serverlessSdk
 
 
-# 1. Initialize SDK instrumentation
-def _get_sdk() -> ServerlessSdk:
-    try:
-        from serverless_aws_lambda_sdk import serverlessSdk
+try:
+    serverlessSdk._initialize()
+except Exception as ex:
+    serverlessSdk._report_error(ex)
 
-        sdk = serverlessSdk
-
-    except ImportError:
-        from .. import serverlessSdk
-
-        sdk = serverlessSdk
-
-    try:
-        sdk._initialize()
-    except Exception as ex:
-        sdk._report_error(ex)
-
-    return sdk
+from ..instrument import Instrumenter  # noqa E402
 
 
-def _check_original_handler_and_reset_vars() -> str:
-    origin = environ.get(Env.ORIGIN_HANDLER)
-
-    if origin:
-        environ[Env.HANDLER] = origin
-        del environ[Env.ORIGIN_HANDLER]
-
-    handler = environ.get(Env.HANDLER)
-
-    if not handler:
-        raise HandlerNotFound(f"{Env.HANDLER} is not set.")
+class HandlerNotFound(Exception):
+    pass
 
 
-def _get_instrumented_handler() -> Handler:
-    _check_original_handler_and_reset_vars()
+class HandlerTypeError(Exception):
+    pass
 
-    handler = _get_handler_function()
-    sdk = _get_sdk()
+
+# To make sure errors appear the same way as uninstrumented
+HandlerNotFound.__name__ = "Runtime.HandlerNotFound"
+HandlerTypeError.__name__ = "TypeError"
+
+
+def _get_instrumented_handler():
+    handler = environ.get("_HANDLER")
+    (module_name, function_name) = handler.rsplit(".", 1)
+    module = importlib.import_module(module_name.replace("/", "."))
+
+    # this is to make sure we report these errors from the invocation phase,
+    # instead of the init phase.
+    def handler_generator():
+        try:
+            _handler = getattr(module, function_name)
+        except AttributeError:
+            raise HandlerNotFound(
+                f"Handler '{function_name}' missing on module '{module_name}'"
+            )
+        if not callable(_handler):
+            raise HandlerTypeError(
+                f"'{type(_handler).__name__}' object is not callable"
+            )
+        return _handler
 
     try:
         instrumenter = Instrumenter()
-        return instrumenter.instrument(handler)
+        return instrumenter.instrument(handler_generator)
     except Exception as ex:
-        sdk._report_error(ex)
-        return handler
+        serverlessSdk._report_error(ex)
+        return handler_generator()
 
 
-handler: Final[Handler] = _get_instrumented_handler()
+handler = _get_instrumented_handler()